Conclude tutorial notebook
