+++ src/springtime/models/phenology_models/__init__.py
@@ -1,0 +1,5 @@
+from .thermaltime import thermaltime
+
+PHENOLOGY_MODELS = {
+    'thermaltime': thermaltime,
+}
