+++ src/springtime/models/phenology_models/phenology_model.py
@@ -1,0 +1,12 @@
+"""Common interface for phenology models."""
+
+from collections.abc import Callable
+from dataclasses import dataclass
+
+
+@dataclass
+class PhenologyModel():
+    predict: Callable
+    params_names: tuple[str, ...]
+    params_defaults: tuple[float, ...]
+    params_bounds: tuple[tuple[float, float], ...]
